move jsonrpc to its own file
Addresses review comment https://github.com/NomicFoundation/hardhat/pull/3383#discussion_r1058480047
